This set of files contains a generalized HBase class that will support CRUD operations on any HBase table. Over the next few days I will check in a cacheing mechanism for HTables as they are expensive to create anew in the spirit of a database connection.  But the interface of this class should not change as this is an internal detail of an HTableDao instance.
First I need to list the prequisites. To get the code here to work you
need to have a working Hadoop and HBase installation. I am using
hbase-0.94.7 and hadoop-1.1.2 and I got them from
http://hadoop.apache.org.  I am using Scala 2.10. 

I am using a MacBook Pro with OS X 10.8.4.  I am running Hadoop as a
pseudo-distributed cluster on my Mac. I have HBase configured to use
HDFS as the file system.  You can read more about how to set up Hadoop
in pseudo-distributed mode in Hadoop, The Definitive Guide.  You can
read about how to configure HBase to use HDFS as the file system in
either Hadoop, The Definitive Guide or HBase, The Definitive Guide.  The
gist of doing this is that you need to start the DFS daemon before you
start the HBase daemon and you need to tell HBase to use HDFS as the
file system. I think its important to configure HBase to use HDFS
because that is how it will work in the real world. Otherwise HBase
stores the data locally on the local file system and you get none of the
scalability and data redundancy that is so important.  

To run this scalatest class you will need to create a table called
'order' with a single column family called 'orderFamily'.  I used the
hbase shell to do this with the following; create 'order', 'orderFamily'